LayoutTests/imported/w3c/web-platform-tests/css/css-text/text-autospace/text-autospace-001-expected.txt

国国XX国
国。XX国
国国XX国
国。XX国
国国11国

FAIL text-autospace-001 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 1 assert_array_equals: expected property 3 to be 45 but got 40 (expected array [40, 40, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 2 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 3 assert_array_equals: expected property 3 to be 45 but got 40 (expected array [40, 40, 40, 45, 40] got [40, 40, 40, 40, 40])
FAIL text-autospace-001 4 assert_array_equals: expected property 1 to be 45 but got 40 (expected array [40, 45, 40, 45, 40] got [40, 40, 40, 40, 40])
